server/requirements.txt
flask
flask-babel
flask-jwt
flask-login
flask-mail
flask-migrate
flask-sqlalchemy
flask-wtf
python-dotenv
